+++ cmd/git/clone-https.go
@@ -1,0 +1,24 @@
+package git
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
+)
+
+// Clone clones a git project into a directory
+func cloneHTTPS(dir, gitUrl, branch string) error {
+
+	_, err := git.PlainClone(dir, false, &git.CloneOptions{
+		URL:           gitUrl,
+		ReferenceName: plumbing.ReferenceName(fmt.Sprintf("refs/heads/%s", branch)),
+		SingleBranch:  true,
+		Progress:      os.Stdout,
+	})
+	if err != nil {
+		return err
+	}
+	return nil
+}
